Update ignore rules creation from specifiers
Only not-equal operator (!=) is relevant.

Also, update tests folder API to reflect the updated Python API,
including the moved utility functions.
Update tests to better reflect edge cases.
